requirements.txt
numpy
pandas
matplotlib
opencv-python
Pillow
torch
torchvision
notebook
scikit-learn
tqdm
